+++ vllm/v1/attention/ops/triton_reshape_and_cache_flash.py
@@ -11,8 +11,8 @@ from vllm.platforms import current_platform
 from vllm.triton_utils import tl, triton
 from vllm.utils.torch_utils import is_quantized_kv_cache
 from vllm.v1.kv_cache_interface import KVQuantMode
-from vllm.v1.attention.ops.fp8e4nv_fp16_sm75 import fp16_to_fp8e4m3
-from vllm.v1.attention.ops.fp8e4nv_sm80 import bf16_to_fp8e4m3
+from vllm.v1.attention.ops.fp8e4nv_bf16 import bf16_to_fp8e4m3
+from vllm.v1.attention.ops.fp8e4nv_fp16 import fp16_to_fp8e4m3
 
 
 def _fp8_software_conv(kv_cache_dtype: str) -> bool:
