+++ vllm/v1/attention/ops/triton_unified_attention.py
@@ -15,8 +15,8 @@ import vllm.envs as envs
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
 from vllm.triton_utils import tl, triton
-from vllm.v1.attention.ops.fp8e4nv_fp16_sm75 import fp8e4m3_to_fp16
-from vllm.v1.attention.ops.fp8e4nv_sm80 import fp8e4m3_to_bf16
+from vllm.v1.attention.ops.fp8e4nv_bf16 import fp8e4m3_to_bf16
+from vllm.v1.attention.ops.fp8e4nv_fp16 import fp8e4m3_to_fp16
 from vllm.v1.attention.ops.triton_attention_helpers import (
     apply_alibi_to_score,
     apply_softcap,
